requirements.txt
Flask>=2.2.2
requests>=2.31.0
gunicorn>=21.2.0
flask_sqlalchemy>=3.1.1
pymysql>=1.1.0
Werkzeug>=2.2.3
python-dotenv==1.0.0